refactor formatBytes function
